Close location list with oustide click, change getCurrentPosition  error response - Close location with oustide element click - Restructure helpers folders - Rename helper component/file - Add react-cool-onclickoutside - Filter displayed error message to user in getCurrentPosition
