+++ package.json
@@ -20,6 +20,7 @@
     "node-sass": "^4.14.1",
     "prettier": "^2.1.2",
     "react": "^16.13.1",
+    "react-cool-onclickoutside": "^1.5.8",
     "react-debounce-input": "^3.2.2",
     "react-dom": "^16.13.1",
     "react-helmet": "^6.1.0",
