+++ src/helpers/components/renderWithContextAndTheme.tsx
@@ -1,11 +1,13 @@
-import React, { FC } from 'react';
+import React, { ComponentType } from 'react';
 import { themeController } from 'theme/theme';
 import { StoreContext } from 'app/App';
 import { ThemeProvider } from 'styled-components';
 import { appState } from 'state/reducers/appReducer';
 
 // Render function with Theme and Context (used for jest testing)
-export const renderWithContextAndTheme = (TestComponent: any): JSX.Element => {
+export const renderWithContextAndTheme = (
+  TestComponent: ComponentType
+): JSX.Element => {
   const state = appState;
   const dispatch = () => {};
   return (
